requirements.txt
streamlit==1.32.2
numpy==1.26.4
tifffile==2024.2.12
Pillow==10.2.0